Lesson 10
Task 1
There is a client.
The client can have any number of cards.
The card has three fields - cvv code, card number and amount on the card.
Display a message indicating how many identical cards the client has.
The field describing the amount on the card should not be included in the comparison.
Those. two multiples with the same cvv codes and numbers, but different amounts are considered the same.
Provide different options for the output message - for example, the client does not have any cards, the client does not have identical cards, and other possible situations.